Revert "sync context"
This reverts commit b1f62e57274b71bbadae2c50a5aad97fc476dd4f.
